1. fix ci 2. address comment
