+++ expression/builtin_other.go
@@ -14,14 +14,8 @@
 package expression
 
 import (
-	"strconv"
-	"strings"
-
 	"github.com/juju/errors"
 	"github.com/pingcap/tidb/context"
-	"github.com/pingcap/tidb/mysql"
-	"github.com/pingcap/tidb/sessionctx/variable"
-	"github.com/pingcap/tidb/terror"
 	"github.com/pingcap/tidb/util/types"
 )
 
@@ -124,562 +118,4 @@ func (b *builtinRowSig) eval(row []types.Datum) (d types.Datum, err error) {
 	}
 	d.SetRow(args)
 	return
-}
-
-type castFunctionClass struct {
-	baseFunctionClass
-
-	tp *types.FieldType
-}
-
-func (c *castFunctionClass) getFunction(args []Expression, ctx context.Context) (builtinFunc, error) {
-	sig := &builtinCastSig{newBaseBuiltinFunc(args, ctx), c.tp}
-	return sig.setSelf(sig), errors.Trace(c.verifyArgs(args))
-}
-
-type builtinCastSig struct {
-	baseBuiltinFunc
-
-	tp *types.FieldType
-}
-
-// eval evals a builtinCastSig.
-// See https://dev.mysql.com/doc/refman/5.7/en/cast-functions.html
-// CastFuncFactory produces builtin function according to field types.
-func (b *builtinCastSig) eval(row []types.Datum) (d types.Datum, err error) {
-	args, err := b.evalArgs(row)
-	if err != nil {
-		return types.Datum{}, errors.Trace(err)
-	}
-	switch b.tp.Tp {
-	// Parser has restricted this.
-	// TypeDouble is used during plan optimization.
-	case mysql.TypeString, mysql.TypeDuration, mysql.TypeDatetime,
-		mysql.TypeDate, mysql.TypeLonglong, mysql.TypeNewDecimal, mysql.TypeDouble:
-		d = args[0]
-		if d.IsNull() {
-			return
-		}
-		return d.ConvertTo(b.ctx.GetSessionVars().StmtCtx, b.tp)
-	}
-	return d, errors.Errorf("unknown cast type - %v", b.tp)
-}
-
-type builtinCastIntAsRealSig struct {
-	baseRealBuiltinFunc
-}
-
-func (b *builtinCastIntAsRealSig) evalReal(row []types.Datum) (res float64, isNull bool, err error) {
-	val, isNull, err := b.args[0].EvalInt(row, b.getCtx().GetSessionVars().StmtCtx)
-	return float64(val), isNull, errors.Trace(err)
-}
-
-type builtinCastIntAsDecimalSig struct {
-	baseDecimalBuiltinFunc
-}
-
-func (b *builtinCastIntAsDecimalSig) evalDecimal(row []types.Datum) (res *types.MyDecimal, isNull bool, err error) {
-	val, isNull, err := b.args[0].EvalInt(row, b.getCtx().GetSessionVars().StmtCtx)
-	return types.NewDecFromInt(val), isNull, errors.Trace(err)
-}
-
-type builtinCastIntAsStringSig struct {
-	baseStringBuiltinFunc
-}
-
-func (b *builtinCastIntAsStringSig) evalString(row []types.Datum) (res string, isNull bool, err error) {
-	val, isNull, err := b.args[0].EvalInt(row, b.getCtx().GetSessionVars().StmtCtx)
-	return strconv.FormatInt(val, 10), isNull, errors.Trace(err)
-}
-
-type builtinCastIntAsTimeSig struct {
-	baseTimeBuiltinFunc
-}
-
-func (b *builtinCastIntAsTimeSig) evalTime(row []types.Datum) (res types.Time, isNull bool, err error) {
-	val, isNull, err := b.args[0].EvalInt(row, b.getCtx().GetSessionVars().StmtCtx)
-	if isNull || err != nil {
-		return types.Time{}, isNull, errors.Trace(err)
-	}
-	res, err = types.ParseTimeFromInt64(val)
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastIntAsDurationSig struct {
-	baseDurationBuiltinFunc
-	fsp int
-}
-
-func (b *builtinCastIntAsDurationSig) evalDuration(row []types.Datum) (res types.Duration, isNull bool, err error) {
-	val, isNull, err := b.args[0].EvalInt(row, b.getCtx().GetSessionVars().StmtCtx)
-	if isNull || err != nil {
-		return types.Duration{}, isNull, errors.Trace(err)
-	}
-	res, err = types.ParseDuration(strconv.FormatInt(val, 10), b.fsp)
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastRealAsIntSig struct {
-	baseIntBuiltinFunc
-}
-
-func (b *builtinCastRealAsIntSig) evalInt(row []types.Datum) (res int64, isNull bool, err error) {
-	val, isNull, err := b.args[0].EvalReal(row, b.getCtx().GetSessionVars().StmtCtx)
-	return int64(val), isNull, errors.Trace(err)
-}
-
-type builtinCastRealAsDecimalSig struct {
-	baseDecimalBuiltinFunc
-}
-
-func (b *builtinCastRealAsDecimalSig) evalDecimal(row []types.Datum) (res *types.MyDecimal, isNull bool, err error) {
-	val, isNull, err := b.args[0].EvalReal(row, b.getCtx().GetSessionVars().StmtCtx)
-	if isNull || err != nil {
-		return nil, isNull, errors.Trace(err)
-	}
-	res = new(types.MyDecimal)
-	err = res.FromFloat64(val)
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastRealAsStringSig struct {
-	baseStringBuiltinFunc
-}
-
-func (b *builtinCastRealAsStringSig) evalString(row []types.Datum) (res string, isNull bool, err error) {
-	val, isNull, err := b.args[0].EvalReal(row, b.getCtx().GetSessionVars().StmtCtx)
-	return strconv.FormatFloat(val, 'f', -1, 64), isNull, errors.Trace(err)
-}
-
-type builtinCastRealAsTimeSig struct {
-	baseTimeBuiltinFunc
-	tp  byte
-	fsp int
-}
-
-func (b *builtinCastRealAsTimeSig) evalTime(row []types.Datum) (res types.Time, isNull bool, err error) {
-	val, isNull, err := b.args[0].EvalReal(row, b.getCtx().GetSessionVars().StmtCtx)
-	res, err = types.ParseTime(strconv.FormatFloat(val, 'f', -1, 64), b.tp, b.fsp)
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastRealAsDurationSig struct {
-	baseDurationBuiltinFunc
-	fsp int
-}
-
-func (b *builtinCastRealAsTimeSig) evalDuration(row []types.Datum) (res types.Duration, isNull bool, err error) {
-	val, isNull, err := b.evalReal(row)
-	if isNull || err != nil {
-		return types.Duration{}, isNull, errors.Trace(err)
-	}
-	res, err = types.ParseDuration(strconv.FormatFloat(val, 'f', -1, 64), b.fsp)
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastDecimalAsIntSig struct {
-	baseIntBuiltinFunc
-}
-
-func (b *builtinCastDecimalAsIntSig) evalInt(row []types.Datum) (res int64, isNull bool, err error) {
-	val, isNull, err := b.args[0].EvalDecimal(row, b.getCtx().GetSessionVars().StmtCtx)
-	if isNull || err != nil {
-		return 0, isNull, errors.Trace(err)
-	}
-	res, err = val.ToInt()
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastDecimalAsStringSig struct {
-	baseStringBuiltinFunc
-}
-
-func (b *builtinCastDecimalAsStringSig) evalString(row []types.Datum) (res string, isNull bool, err error) {
-	val, isNull, err := b.args[0].EvalDecimal(row, b.getCtx().GetSessionVars().StmtCtx)
-	if isNull || err != nil {
-		return "", isNull, errors.Trace(err)
-	}
-	return string(val.ToString()), false, nil
-}
-
-type builtinCastDecimalAsRealSig struct {
-	baseRealBuiltinFunc
-}
-
-func (b *builtinCastDecimalAsRealSig) evalReal(row []types.Datum) (res float64, isNull bool, err error) {
-	val, isNull, err := b.args[0].EvalDecimal(row, b.getCtx().GetSessionVars().StmtCtx)
-	if isNull || err != nil {
-		return 0, isNull, errors.Trace(err)
-	}
-	res, err = val.ToFloat64()
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastDecimalAsTimeSig struct {
-	baseTimeBuiltinFunc
-	tp byte
-}
-
-func (b *builtinCastDecimalAsTimeSig) evalTime(row []types.Datum) (res types.Time, isNull bool, err error) {
-	val, isNull, err := b.evalDecimal(row)
-	if isNull || err != nil {
-		return types.Time{}, isNull, errors.Trace(err)
-	}
-	_, fsp := val.PrecisionAndFrac()
-	res, err = types.ParseTime(string(val.ToString()), b.tp, fsp)
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastDecimalAsDurationSig struct {
-	baseDurationBuiltinFunc
-}
-
-func (b *builtinCastDecimalAsDurationSig) evalDuration(row []types.Datum) (res types.Duration, isNull bool, err error) {
-	val, isNull, err := b.evalDecimal(row)
-	if isNull || err != nil {
-		return types.Duration{}, false, errors.Trace(err)
-	}
-	_, fsp := val.PrecisionAndFrac()
-	res, err = types.ParseDuration(string(val.ToString()), fsp)
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastStringAsIntSig struct {
-	baseIntBuiltinFunc
-}
-
-func (b *builtinCastStringAsIntSig) evalInt(row []types.Datum) (res int64, isNull bool, err error) {
-	if types.IsHybridType(b.args[0].GetType().Tp) {
-		return b.args[0].EvalInt(row, b.getCtx().GetSessionVars().StmtCtx)
-	}
-	val, isNull, err := b.args[0].EvalString(row, b.getCtx().GetSessionVars().StmtCtx)
-	if isNull || err != nil {
-		return 0, isNull, errors.Trace(err)
-	}
-	res, err = strconv.ParseInt(val, 10, 64)
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastStringAsRealSig struct {
-	baseRealBuiltinFunc
-}
-
-func (b *builtinCastStringAsRealSig) evalReal(row []types.Datum) (res float64, isNull bool, err error) {
-	if types.IsHybridType(b.args[0].GetType().Tp) {
-		return b.args[0].EvalReal(row, b.getCtx().GetSessionVars().StmtCtx)
-	}
-	val, isNull, err := b.args[0].EvalString(row, b.getCtx().GetSessionVars().StmtCtx)
-	if isNull || err != nil {
-		return 0, isNull, errors.Trace(err)
-	}
-	res, err = strconv.ParseFloat(val, 64)
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastStringAsDecimalSig struct {
-	baseDecimalBuiltinFunc
-}
-
-func (b *builtinCastStringAsDecimalSig) evalDecimal(row []types.Datum) (res *types.MyDecimal, isNull bool, err error) {
-	if types.IsHybridType(b.args[0].GetType().Tp) {
-		return b.args[0].EvalDecimal(row, b.getCtx().GetSessionVars().StmtCtx)
-	}
-	val, isNull, err := b.args[0].EvalString(row, b.getCtx().GetSessionVars().StmtCtx)
-	if isNull || err != nil {
-		return nil, isNull, errors.Trace(err)
-	}
-	res = new(types.MyDecimal)
-	err = res.FromString([]byte(val))
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastStringAsTimeSig struct {
-	baseTimeBuiltinFunc
-	tp  byte
-	fsp int
-}
-
-func (b *builtinCastStringAsTimeSig) evalTime(row []types.Datum) (res types.Time, isNull bool, err error) {
-	val, isNull, err := b.evalString(row)
-	if isNull || err != nil {
-		return types.Time{}, isNull, errors.Trace(err)
-	}
-	res, err = types.ParseTime(val, b.tp, b.fsp)
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastStringAsDurationSig struct {
-	baseDurationBuiltinFunc
-	fsp int
-}
-
-func (b *builtinCastStringAsDurationSig) evalDuration(row []types.Datum) (res types.Duration, isNull bool, err error) {
-	val, isNull, err := b.evalString(row)
-	if isNull || err != nil {
-		return types.Duration{}, isNull, errors.Trace(err)
-	}
-	res, err = types.ParseDuration(val, b.fsp)
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastTimeAsIntSig struct {
-	baseIntBuiltinFunc
-}
-
-func (b *builtinCastTimeAsIntSig) evalInt(row []types.Datum) (res int64, isNull bool, err error) {
-	val, isNull, err := b.evalTime(row)
-	if isNull || err != nil {
-		return 0, isNull, errors.Trace(err)
-	}
-	res, err = val.ToNumber().ToInt()
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastTimeAsRealSig struct {
-	baseRealBuiltinFunc
-}
-
-func (b *builtinCastTimeAsRealSig) evalReal(row []types.Datum) (res float64, isNull bool, err error) {
-	val, isNull, err := b.evalTime(row)
-	if isNull || err != nil {
-		return 0, isNull, errors.Trace(err)
-	}
-	res, err = val.ToNumber().ToFloat64()
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastTimeAsDecimalSig struct {
-	baseDecimalBuiltinFunc
-}
-
-func (b *builtinCastTimeAsDecimalSig) evalDecimal(row []types.Datum) (res *types.MyDecimal, isNull bool, err error) {
-	val, isNull, err := b.evalTime(row)
-	if isNull || err != nil {
-		return nil, isNull, errors.Trace(err)
-	}
-	return val.ToNumber(), false, errors.Trace(err)
-}
-
-type builtinCastTimeAsStringSig struct {
-	baseStringBuiltinFunc
-}
-
-func (b *builtinCastTimeAsStringSig) evalString(row []types.Datum) (res string, isNull bool, err error) {
-	val, isNull, err := b.evalTime(row)
-	if isNull || err != nil {
-		return "", isNull, errors.Trace(err)
-	}
-	return val.String(), false, errors.Trace(err)
-}
-
-type builtinCastTimeAsDurationSig struct {
-	baseDurationBuiltinFunc
-}
-
-func (b *builtinCastTimeAsDurationSig) evalDuration(row []types.Datum) (res types.Duration, isNull bool, err error) {
-	val, isNull, err := b.evalTime(row)
-	if isNull || err != nil {
-		return types.Duration{}, isNull, errors.Trace(err)
-	}
-	res, err = val.ConvertToDuration()
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastDurationAsIntSig struct {
-	baseIntBuiltinFunc
-}
-
-func (b *builtinCastDurationAsIntSig) evalInt(row []types.Datum) (res int64, isNull bool, err error) {
-	val, isNull, err := b.evalDuration(row)
-	if isNull || err != nil {
-		return 0, isNull, errors.Trace(err)
-	}
-	res, err = val.ToNumber().ToInt()
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastDurationAsRealSig struct {
-	baseRealBuiltinFunc
-}
-
-func (b *builtinCastDurationAsRealSig) evalReal(row []types.Datum) (res float64, isNull bool, err error) {
-	val, isNull, err := b.evalDuration(row)
-	if isNull || err != nil {
-		return 0, isNull, errors.Trace(err)
-	}
-	res, err = val.ToNumber().ToFloat64()
-	return res, false, errors.Trace(err)
-}
-
-type builtinCastDurationAsDecimalSig struct {
-	baseDecimalBuiltinFunc
-}
-
-func (b *builtinCastDurationAsDecimalSig) evalDecimal(row []types.Datum) (res *types.MyDecimal, isNull bool, err error) {
-	val, isNull, err := b.evalDuration(row)
-	if isNull || err != nil {
-		return nil, isNull, errors.Trace(err)
-	}
-	return val.ToNumber(), false, errors.Trace(err)
-}
-
-type builtinCastDurationAsStringSig struct {
-	baseStringBuiltinFunc
-}
-
-func (b *builtinCastDurationAsStringSig) evalString(row []types.Datum) (res string, isNull bool, err error) {
-	val, isNull, err := b.evalDuration(row)
-	if isNull || err != nil {
-		return "", isNull, errors.Trace(err)
-	}
-	return val.String(), false, errors.Trace(err)
-}
-
-type builtinCastDurationAsTimeSig struct {
-	baseTimeBuiltinFunc
-	tp byte
-}
-
-func (b *builtinCastDurationAsTimeSig) evalTime(row []types.Datum) (res types.Time, isNull bool, err error) {
-	val, isNull, err := b.evalDuration(row)
-	if isNull || err != nil {
-		return types.Time{}, isNull, errors.Trace(err)
-	}
-	res, err = val.ConvertToTime(b.tp)
-	return res, false, errors.Trace(err)
-}
-
-type setVarFunctionClass struct {
-	baseFunctionClass
-}
-
-func (c *setVarFunctionClass) getFunction(args []Expression, ctx context.Context) (builtinFunc, error) {
-	err := errors.Trace(c.verifyArgs(args))
-	bt := &builtinSetVarSig{newBaseBuiltinFunc(args, ctx)}
-	bt.deterministic = false
-	return bt.setSelf(bt), errors.Trace(err)
-}
-
-type builtinSetVarSig struct {
-	baseBuiltinFunc
-}
-
-func (b *builtinSetVarSig) eval(row []types.Datum) (types.Datum, error) {
-	args, err := b.evalArgs(row)
-	if err != nil {
-		return types.Datum{}, errors.Trace(err)
-	}
-	sessionVars := b.ctx.GetSessionVars()
-	varName, _ := args[0].ToString()
-	if !args[1].IsNull() {
-		strVal, err := args[1].ToString()
-		if err != nil {
-			return types.Datum{}, errors.Trace(err)
-		}
-		sessionVars.Users[varName] = strings.ToLower(strVal)
-	}
-	return args[1], nil
-}
-
-type getVarFunctionClass struct {
-	baseFunctionClass
-}
-
-func (c *getVarFunctionClass) getFunction(args []Expression, ctx context.Context) (builtinFunc, error) {
-	err := errors.Trace(c.verifyArgs(args))
-	bt := &builtinGetVarSig{newBaseBuiltinFunc(args, ctx)}
-	bt.deterministic = false
-	return bt.setSelf(bt), errors.Trace(err)
-}
-
-type builtinGetVarSig struct {
-	baseBuiltinFunc
-}
-
-func (b *builtinGetVarSig) eval(row []types.Datum) (types.Datum, error) {
-	args, err := b.evalArgs(row)
-	if err != nil {
-		return types.Datum{}, errors.Trace(err)
-	}
-	sessionVars := b.ctx.GetSessionVars()
-	varName, _ := args[0].ToString()
-	if v, ok := sessionVars.Users[varName]; ok {
-		return types.NewDatum(v), nil
-	}
-	return types.Datum{}, nil
-}
-
-type valuesFunctionClass struct {
-	baseFunctionClass
-
-	offset int
-}
-
-func (c *valuesFunctionClass) getFunction(args []Expression, ctx context.Context) (builtinFunc, error) {
-	err := errors.Trace(c.verifyArgs(args))
-	bt := &builtinValuesSig{newBaseBuiltinFunc(args, ctx), c.offset}
-	bt.deterministic = false
-	return bt.setSelf(bt), errors.Trace(err)
-}
-
-type builtinValuesSig struct {
-	baseBuiltinFunc
-
-	offset int
-}
-
-func (b *builtinValuesSig) eval(_ []types.Datum) (types.Datum, error) {
-	values := b.ctx.GetSessionVars().CurrInsertValues
-	if values == nil {
-		return types.Datum{}, errors.New("Session current insert values is nil")
-	}
-	row := values.([]types.Datum)
-	if len(row) > b.offset {
-		return row[b.offset], nil
-	}
-	return types.Datum{}, errors.Errorf("Session current insert values len %d and column's offset %v don't match", len(row), b.offset)
-}
-
-type bitCountFunctionClass struct {
-	baseFunctionClass
-}
-
-func (c *bitCountFunctionClass) getFunction(args []Expression, ctx context.Context) (builtinFunc, error) {
-	sig := &builtinBitCountSig{newBaseBuiltinFunc(args, ctx)}
-	return sig.setSelf(sig), errors.Trace(c.verifyArgs(args))
-}
-
-type builtinBitCountSig struct {
-	baseBuiltinFunc
-}
-
-// eval evals a builtinBitCountSig.
-// See https://dev.mysql.com/doc/refman/5.7/en/bit-functions.html#function_bit-count
-func (b *builtinBitCountSig) eval(row []types.Datum) (d types.Datum, err error) {
-	args, err := b.evalArgs(row)
-	if err != nil {
-		return d, errors.Trace(err)
-	}
-	arg := args[0]
-	if arg.IsNull() {
-		return d, nil
-	}
-	sc := new(variable.StatementContext)
-	sc.IgnoreTruncate = true
-	bin, err := arg.ToInt64(sc)
-	if err != nil {
-		if terror.ErrorEqual(err, types.ErrOverflow) {
-			d.SetInt64(64)
-			return d, nil
-
-		}
-		return d, errors.Trace(err)
-	}
-	var count int64
-	for bin != 0 {
-		count++
-		bin = (bin - 1) & bin
-	}
-	d.SetInt64(count)
-	return d, nil
 }
